--- cookiecutter.json
@@ -1,27 +1,0 @@
-{
-    "full_name": "John Smith",
-    "project_name": "My Research Project",
-    "project_slug": "{{ cookiecutter.project_name.lower().replace(' ', '_').replace('-', '_') }}",
-    "project_short_description": "Research repository with an installable package.",
-    "line_length": "120",
-    "python_version": 3.11,
-    "mypy": false,
-    "open_source_license": [
-        "MIT license",
-        "BSD license",
-        "ISC license",
-        "Apache Software License 2.0",
-        "GNU General Public License v3",
-        "Not open source"
-    ],
-    "__prompts__": {
-        "full_name": "Type your name:",
-        "project_name": "Give your project a name:",
-        "project_slug": "Choose a name for your installable Python package:",
-        "project_short_description": "Provide a short description of your project:",
-        "line_length": "Choose a maximum line length for code formatting:",
-        "python_version": "Choose a version of Python:",
-        "mypy": "Would you like to use mypy for static type checking?",
-        "open_source_license": "Choose a license:"
-    }
-}
