+++ ruff.toml
@@ -1,0 +1,4 @@
+# Exclude template which isn't parseable.
+exclude = [
+  '*cookiecutter.project_slug*'
+]
