--- {{cookiecutter.project_slug}}/.python-version
@@ -1,1 +1,0 @@
-{{cookiecutter.python_version}}
