--- {{cookiecutter.project_slug}}/src/{{cookiecutter.project_slug}}/__init__.py
@@ -1,1 +1,0 @@
-"""Top-level package for {{ cookiecutter.project_name }}."""
