+++ {{project_slug}}/.editorconfig
