+++ {{project_slug}}/.gitignore
