+++ {{project_slug}}/.pre-commit-config.yaml.jinja
@@ -26,7 +26,7 @@ repos:
       require_serial: true
       language: python
       types_or: [python]
-{% if cookiecutter.mypy -%}
+{% if mypy -%}
 - repo: local
   hooks:
 # Static type-checking.
