+++ {{project_slug}}/.python-version.jinja
@@ -1,0 +1,1 @@
+{{python_version}}
