+++ {{project_slug}}/LICENSE.jinja
@@ -1,7 +1,7 @@
-{% if cookiecutter.open_source_license == 'MIT license' -%}
+{% if open_source_license == 'MIT license' -%}
 MIT License
 
-Copyright (c) {% now 'local', '%Y' %}, {{ cookiecutter.full_name }}
+Copyright (c) {{ current_year }}, {{ full_name }}
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -20,11 +20,11 @@ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
-{% elif cookiecutter.open_source_license == 'BSD license' %}
+{% elif open_source_license == 'BSD license' %}
 
 BSD License
 
-Copyright (c) {% now 'local', '%Y' %}, {{ cookiecutter.full_name }}
+Copyright (c) {{ current_year }}, {{ full_name }}
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification,
@@ -51,18 +51,18 @@ DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
 OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
 OF THE POSSIBILITY OF SUCH DAMAGE.
-{% elif cookiecutter.open_source_license == 'ISC license' -%}
+{% elif open_source_license == 'ISC license' -%}
 ISC License
 
-Copyright (c) {% now 'local', '%Y' %}, {{ cookiecutter.full_name }}
+Copyright (c) {{ current_year }}, {{ full_name }}
 
 Permission to use, copy, modify, and/or distribute this software for any purpose with or without fee is hereby granted, provided that the above copyright notice and this permission notice appear in all copies.
 
 THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
-{% elif cookiecutter.open_source_license == 'Apache Software License 2.0' -%}
+{% elif open_source_license == 'Apache Software License 2.0' -%}
 Apache Software License 2.0
 
-Copyright (c) {% now 'local', '%Y' %}, {{ cookiecutter.full_name }}
+Copyright (c) {{ current_year }}, {{ full_name }}
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -75,12 +75,12 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
-{% elif cookiecutter.open_source_license == 'GNU General Public License v3' -%}
+{% elif open_source_license == 'GNU General Public License v3' -%}
 GNU GENERAL PUBLIC LICENSE
-                      Version 3, 29 June 2007
+Version 3, 29 June 2007
 
-    {{ cookiecutter.project_short_description }}
-    Copyright (C) {% now 'local', '%Y' %}  {{ cookiecutter.full_name }}
+{{ project_short_description }}
+Copyright (C) {{ current_year }}  {{ full_name }}
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
