+++ {{project_slug}}/MANIFEST.in
