+++ {{project_slug}}/README.md.jinja
@@ -1,7 +1,7 @@
-{{ cookiecutter.project_name }}
-{% for _ in cookiecutter.project_name %}={% endfor %}
+{{ project_name }}
+{% for _ in project_name %}={% endfor %}
 
-{{ cookiecutter.project_short_description }}
+{{ project_short_description }}
 
 
 Environment set-up
