+++ {{project_slug}}/scripts/config/main.yaml
