+++ {{project_slug}}/scripts/example.py.jinja
@@ -6,7 +6,7 @@ from dotenv import load_dotenv
 import hydra
 import os
 import logging
-from {{cookiecutter.project_slug}}.utils import func, func2
+from {{project_slug}}.utils import func, func2
 
 
 logger = logging.getLogger(__name__) # Hydra configures logging for us.
