+++ {{project_slug}}/src/{{project_slug}}/__init__.py.jinja
@@ -1,0 +1,1 @@
+"""Top-level package for {{ project_name }}."""
