+++ {{project_slug}}/tests/__init__.py.jinja
@@ -1,0 +1,1 @@
+"""Unit test package for {{ project_slug }}."""
