+++ {{project_slug}}/tests/test_utils.py.jinja
@@ -6,7 +6,7 @@ import os
 import numpy as np
 import pytest
 
-from {{cookiecutter.project_slug}}.utils import func, func2, patchify_images
+from {{project_slug}}.utils import func, func2, patchify_images
 
 @pytest.fixture(scope='function')  # https://docs.pytest.org/en/6.2.x/fixture.html#fixture-scopes
 def my_fixture():
